requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
dateparser==1.2.0
Deprecated==1.2.18
Flask==2.3.3
frozenlist==1.6.0
future==1.0.0
google-api-core==2.24.1
google-api-python-client==2.108.0
google-auth==2.40.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
ja_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.7.0/ja_core_news_sm-3.7.0.tar.gz#sha256=5cb1c87cd0551404a03fd630e824dfcb32e793f2ae5331a196d6e346749bdb2d
jaconv==0.3.4
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
line-bot-sdk==3.7.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
multidict==6.4.3
murmurhash==1.0.12
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==25.0
preshed==3.0.9
propcache==0.3.1
proto-plus==1.26.1
protobuf>=3.19.5,<6.0.0.dev0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250129
SudachiPy==0.6.10
tenacity==9.1.2
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
redis
flask-session
flask-limiter
